# server/src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Record_string.string_": {
				"properties": {},
				"additionalProperties": {
					"type": "string"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"ApiResponse_any-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {},
						"type": "array",
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserRole": {
				"type": "string",
				"enum": [
					"ADMIN",
					"PROJECT_MANAGER",
					"COLLABORATOR"
				]
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"mustResetPassword": {
						"type": "boolean"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"updatedAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"required": [
					"id",
					"email",
					"name",
					"role",
					"mustResetPassword",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserProfile": {
				"$ref": "#/components/schemas/User"
			},
			"ApiResponse_UserProfile_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserProfile"
							}
						],
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProfileRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [],
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChangePasswordRequest": {
				"properties": {
					"currentPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					},
					"confirmNewPassword": {
						"type": "string"
					}
				},
				"required": [
					"currentPassword",
					"newPassword",
					"confirmNewPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_36_Enums.Priority": {
				"type": "string",
				"enum": [
					"LOW",
					"MEDIUM",
					"HIGH"
				]
			},
			"_36_Enums.Status": {
				"type": "string",
				"enum": [
					"TODO",
					"IN_PROGRESS",
					"COMPLETED"
				]
			},
			"DefaultSelection_Prisma._36_TaskPayload_": {
				"properties": {
					"projectId": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"position": {
						"type": "number",
						"format": "double"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.Status"
					},
					"priority": {
						"$ref": "#/components/schemas/_36_Enums.Priority"
					},
					"dueDate": {
						"type": "string",
						"format": "date-time"
					},
					"description": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"projectId",
					"updatedAt",
					"createdAt",
					"position",
					"tags",
					"status",
					"priority",
					"dueDate",
					"description",
					"title",
					"id"
				],
				"type": "object"
			},
			"Task": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_TaskPayload_",
				"description": "Model Task"
			},
			"ApiResponse_Task-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Task"
						},
						"type": "array",
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommentAuthor": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					}
				},
				"required": [
					"id",
					"email",
					"name",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Attachment": {
				"properties": {
					"id": {
						"type": "string"
					},
					"filename": {
						"type": "string"
					},
					"fileKey": {
						"type": "string"
					},
					"presignedUrl": {
						"type": "string"
					},
					"mimeType": {
						"type": "string"
					},
					"fileSize": {
						"type": "number",
						"format": "double"
					},
					"taskId": {
						"type": "string"
					},
					"commentId": {
						"type": "string",
						"nullable": true
					},
					"uploadedById": {
						"type": "string"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"required": [
					"id",
					"filename",
					"fileKey",
					"mimeType",
					"fileSize",
					"taskId",
					"commentId",
					"uploadedById",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Comment": {
				"properties": {
					"id": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"taskId": {
						"type": "string"
					},
					"authorId": {
						"type": "string"
					},
					"author": {
						"$ref": "#/components/schemas/CommentAuthor"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"updatedAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"attachments": {
						"items": {
							"$ref": "#/components/schemas/Attachment"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"content",
					"taskId",
					"authorId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TaskAssignee": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string"
					},
					"assignedAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"required": [
					"userId",
					"name",
					"email",
					"assignedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SharedTask": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"dueDate": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"priority": {
						"type": "string",
						"enum": [
							"LOW",
							"MEDIUM",
							"HIGH"
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"TODO",
							"IN_PROGRESS",
							"COMPLETED"
						]
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"position": {
						"type": "number",
						"format": "double"
					},
					"projectId": {
						"type": "string"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"updatedAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"comments": {
						"items": {
							"$ref": "#/components/schemas/Comment"
						},
						"type": "array"
					},
					"attachments": {
						"items": {
							"$ref": "#/components/schemas/Attachment"
						},
						"type": "array"
					},
					"assignees": {
						"items": {
							"$ref": "#/components/schemas/TaskAssignee"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"title",
					"priority",
					"status",
					"tags",
					"position",
					"projectId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_SharedTask_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/SharedTask"
							}
						],
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Task_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Task"
							}
						],
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTaskRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"projectId": {
						"type": "string"
					},
					"dueDate": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"priority": {
						"type": "string",
						"enum": [
							"LOW",
							"MEDIUM",
							"HIGH"
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"TODO",
							"IN_PROGRESS",
							"COMPLETED"
						]
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"position": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"projectId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTaskRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"dueDate": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"priority": {
						"type": "string",
						"enum": [
							"LOW",
							"MEDIUM",
							"HIGH"
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"TODO",
							"IN_PROGRESS",
							"COMPLETED"
						]
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"position": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"DefaultSelection_Prisma._36_TaskAssignmentPayload_": {
				"properties": {
					"assignedAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "string"
					},
					"taskId": {
						"type": "string"
					}
				},
				"required": [
					"assignedAt",
					"userId",
					"taskId"
				],
				"type": "object"
			},
			"TaskAssignment": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_TaskAssignmentPayload_",
				"description": "Model TaskAssignment"
			},
			"ApiResponse_TaskAssignment_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/TaskAssignment"
							}
						],
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AssignUserInput": {
				"properties": {
					"userId": {
						"type": "string"
					}
				},
				"required": [
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TaskAssignment-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/TaskAssignment"
						},
						"type": "array",
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BulkAssignInput": {
				"properties": {
					"userIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"userIds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TaskAssignee-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/TaskAssignee"
						},
						"type": "array",
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "@nextask/server",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
		"/users/search/autocomplete": {
			"get": {
				"operationId": "GetUserAutocomplete",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any-Array_"
								}
							}
						}
					}
				},
				"description": "Returns a list of team members for autocomplete search dropdowns (Task 3.4)",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"project:member"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserProfile_"
								}
							}
						}
					}
				},
				"description": "Returns the authenticated user's profile.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"patch": {
				"operationId": "UpdateMe",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserProfile_"
								}
							}
						}
					}
				},
				"description": "Updates the authenticated user's name and/or email.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProfileRequest"
							}
						}
					}
				}
			}
		},
		"/users/me/change-password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Allows an authenticated, onboarded user to change their password voluntarily.\r\n(First-login forced reset uses POST /auth/reset-password instead.)",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordRequest"
							}
						}
					}
				}
			}
		},
		"/users/me/projects": {
			"get": {
				"operationId": "GetMyProjects",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any-Array_"
								}
							}
						}
					}
				},
				"description": "Returns a list of projects the authenticated user belongs to.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/tasks": {
			"get": {
				"operationId": "GetTasks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Task-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": [
							"project:member"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"TODO",
								"IN_PROGRESS",
								"COMPLETED"
							]
						}
					},
					{
						"in": "query",
						"name": "priority",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"LOW",
								"MEDIUM",
								"HIGH"
							]
						}
					},
					{
						"in": "query",
						"name": "tags",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					}
				]
			},
			"post": {
				"operationId": "CreateNewTask",
				"responses": {
					"201": {
						"description": "Task Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Task_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": [
							"project:manager"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTaskRequest"
							}
						}
					}
				}
			}
		},
		"/tasks/{id}": {
			"get": {
				"operationId": "GetTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_SharedTask_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": [
							"project:member"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateExistingTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Task_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": [
							"project:member"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTaskRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteExistingTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": [
							"project:manager"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/tasks/{taskId}/assignees": {
			"post": {
				"operationId": "AssignUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TaskAssignment_"
								}
							}
						}
					}
				},
				"description": "Assigns a user to a task.",
				"tags": [
					"Task Assignments"
				],
				"security": [
					{
						"jwt": [
							"project:manager"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AssignUserInput"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "BulkAssign",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TaskAssignment-Array_"
								}
							}
						}
					}
				},
				"description": "Bulk assigns users to a task, overwriting existing assignments.",
				"tags": [
					"Task Assignments"
				],
				"security": [
					{
						"jwt": [
							"project:manager"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BulkAssignInput"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAssignees",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TaskAssignee-Array_"
								}
							}
						}
					}
				},
				"description": "Retrieves all assignees for a given task.",
				"tags": [
					"Task Assignments"
				],
				"security": [
					{
						"jwt": [
							"project:member"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/tasks/{taskId}/assignees/{userId}": {
			"delete": {
				"operationId": "UnassignUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Unassigns a user from a task.",
				"tags": [
					"Task Assignments"
				],
				"security": [
					{
						"jwt": [
							"project:manager"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}